Refactor watershed map god component.
